Set OUTPUT mode after writing LOW
When pullups are used, this disables the pullups before switching to
OUTPUT LOW. Previously, this would switch to OUTPUT HIGH briefly before
going to OUTPUT LOW, violating the I²C spec and possibly damaging
slaves. Now, the pullups are briefly disabled, which isn't great either,
but most AVR hardware has no way around this.

When pullups are disabled, this does not change anything.