Extend support for non-AVR architectures
There is no generic Arduino function to read the current interrupt
status, only to enable and disable interrupts. For AVR architectures
ATOMIC_BLOCK is used to prevent interrupts occuring when the the SDA or
SCL pin directions are changed, which could result in spurious signals
on the I2C bus. Other architectures do not have such protection.